Sorted commands into separate directory
